(feat/tasks): refactor lint_rules for subsequent todos
